Add nrelatb tests and move test over from generation_test
Add pytest-mock dependency --used in the new tests -- to environment.yml and pyproject.toml
