+++ tests/generation_test.py
@@ -25,7 +25,7 @@ from powergenome.GenX import (
     round_col_values,
     set_int_cols,
 )
-from powergenome.nrelatb import db_col_values, regional_capex_multiplier
+from powergenome.nrelatb import db_col_values
 
 CWD = Path.cwd()
 # os.environ["RESOURCE_GROUPS"] = str(CWD / "data" / "resource_groups_base")
@@ -1429,120 +1429,3 @@ class TestLabelRetirementYear:
 
         # Check if the original dataframe is returned without any changes
         assert df.equals(labeled_df)
-
-
-class TestRegionalCapexMultiplier:
-    """Test class for regional_capex_multiplier function."""
-
-    @pytest.fixture(scope="class")
-    def sample_data(self):
-        """Create a sample DataFrame with different technology names, including special regex characters."""
-        data = {
-            "technology": ["Solar PV", "Wind*", "Battery+", "CCGT (new)."],
-            "Inv_Cost_per_MWyr": [1000, 2000, 3000, 4000],
-            "Inv_Cost_per_MWhyr": [500, 1000, 1500, 2000],
-        }
-        return pd.DataFrame(data)
-
-    @pytest.fixture(scope="class")
-    def region_map(self):
-        """Create a mapping of regions to cost adjustment regions."""
-        return {"Region1": "CostRegionA", "Region2": "CostRegionB"}
-
-    @pytest.fixture(scope="class")
-    def tech_map(self):
-        """Create a mapping of technology names to their EIA equivalents."""
-        return {
-            "Solar PV": "Solar PV",
-            "Wind*": "Wind",
-            "Battery+": "Battery Storage",
-            "CCGT (new).": "Combined Cycle",
-        }
-
-    @pytest.fixture(scope="class")
-    def regional_multipliers(self):
-        """Create a DataFrame of regional multipliers."""
-        data = {
-            "CostRegionA": {
-                "Solar PV": 1.1,
-                "Wind": 1.2,
-                "Battery Storage": 0.9,
-                "Combined Cycle": 1.05,
-            },
-            "CostRegionB": {
-                "Solar PV": 1.0,
-                "Wind": 1.1,
-                "Battery Storage": 0.95,
-                "Combined Cycle": 1.02,
-            },
-        }
-        return pd.DataFrame(data).T
-
-    def test_regional_capex_multiplier(
-        self, sample_data, region_map, tech_map, regional_multipliers
-    ):
-        """Test that regional multipliers are applied correctly, including for technologies with special regex characters."""
-        region = "Region1"
-        modified_df = regional_capex_multiplier(
-            sample_data, region, region_map, tech_map, regional_multipliers
-        )
-
-        expected_inv_cost = [1000 * 1.1, 2000 * 1.2, 3000 * 0.9, 4000 * 1.05]
-        expected_inv_cost_mwh = [500 * 1.1, 1000 * 1.2, 1500 * 0.9, 2000 * 1.05]
-
-        assert modified_df["Inv_Cost_per_MWyr"].tolist() == pytest.approx(
-            expected_inv_cost, rel=1e-3
-        )
-        assert modified_df["Inv_Cost_per_MWhyr"].tolist() == pytest.approx(
-            expected_inv_cost_mwh, rel=1e-3
-        )
-
-    def test_missing_technology_in_multipliers(self, sample_data, region_map, tech_map):
-        """Test that missing technologies in the multipliers default to the average multiplier."""
-        region = "Region1"
-        # Modify multipliers so "CCGT (new)." is missing
-        regional_multipliers = pd.DataFrame(
-            {
-                "CostRegionA": {
-                    "Solar PV": 1.1,
-                    "Wind": 1.2,
-                    "Battery Storage": 0.9,
-                    "Combined Cycle": np.nan,
-                }
-            }
-        ).T
-
-        modified_df = regional_capex_multiplier(
-            sample_data, region, region_map, tech_map, regional_multipliers
-        )
-
-        # The "CCGT (new)." should use the average multiplier (1.067)
-        avg_multiplier = sum([1.1, 1.2, 0.9]) / 3
-        expected_inv_cost = [1000 * 1.1, 2000 * 1.2, 3000 * 0.9, 4000 * avg_multiplier]
-
-        assert modified_df["Inv_Cost_per_MWyr"].tolist() == pytest.approx(
-            expected_inv_cost, rel=1e-3
-        )
-
-    def test_empty_dataframe(self, region_map, tech_map, regional_multipliers):
-        """Test handling of an empty input DataFrame."""
-        empty_df = pd.DataFrame(
-            columns=["technology", "Inv_Cost_per_MWyr", "Inv_Cost_per_MWhyr"]
-        )
-        region = "Region1"
-
-        modified_df = regional_capex_multiplier(
-            empty_df, region, region_map, tech_map, regional_multipliers
-        )
-        assert modified_df.empty
-
-    def test_invalid_region(
-        self, sample_data, region_map, tech_map, regional_multipliers
-    ):
-        """Test behavior when an invalid region is provided."""
-        invalid_region = "UnknownRegion"
-
-        with pytest.raises(KeyError):
-            regional_capex_multiplier(
-                sample_data, invalid_region, region_map, tech_map, regional_multipliers
-            )
